// fem-vision-studio/design-mockup/src/lib/types/concrete/MnKappaResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { InteractionPoint } from "./InteractionPoint";
import type { MnKappaDiagram } from "./MnKappaDiagram";

/**
 * Antwoord op [`crate::MnKappaRequest`].
 */
export type MnKappaResponse = { section_name: string, reinforcement_summary: string, f_cd_mpa: number, f_yd_mpa: number, d_mm: number, a_s_bottom_mm2: number, a_s_top_mm2: number, 
/**
 * Drukcapaciteit N_Rd,c (kN, positief getal) en trekcapaciteit N_Rd,t.
 */
n_rd_compression_kn: number, n_rd_tension_kn: number, diagram: MnKappaDiagram, interaction_positive: Array<InteractionPoint>, interaction_negative: Array<InteractionPoint>, };